MySQLPreparedStatment first test for select works. need to move output buffers to result set. need to implement other get methods. Need to add typedef for buffer vectors.

+++ include/dbccpp/PreparedStatement.h
@@ -191,7 +191,7 @@ public:
      */
     // FIXME: this should be a 64-bit type really
     // FIXME: information is SQLite-specific
-    virtual int getLastInsertId() = 0;
+    virtual u_int64_t getLastInsertId() = 0;
 
     /** Get the underlying SQL statement. */
     virtual const char* getSQL() const = 0;
